requirements.txt
tensorflow
keras
gym
matplotlib
numpy
pandas
google-cloud
google-generativeai
PyYAML
python-dotenv
paramiko
